+++ R/RcppModule.R
@@ -1,3 +1,4 @@
 #' @useDynLib nmslibR, .registration = TRUE
 #' @importFrom Rcpp evalCpp
+#' @importFrom lifecycle deprecate_warn is_present
 NULL