+++ R/package.R
@@ -1,12 +1,16 @@
-#-------------------------
-# Load the python-modules
-#-------------------------
+#---------------------------------------------------------------------------------
+# An alternative way of configuration - not tested on CRAN yet - is the following:
+#   https://github.com/rstudio/reticulate/issues/883#issuecomment-775552812
+#   https://github.com/kevinushey/usespandas/blob/master/DESCRIPTION
+#   https://github.com/kevinushey/usespandas/blob/master/R/zzz.R
+#---------------------------------------------------------------------------------
 
 
 NMSLIB <- NULL; SCP <- NULL;
 
-
 .onLoad <- function(libname, pkgname) {
+
+  # reticulate::configure_environment(pkgname, force = TRUE)               # this R programming line is related to the weblinks at the top of the file (see also the documentation)
 
   try({
     if (reticulate::py_available(initialize = FALSE)) {